+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Added ViewConstants StructuredBuffer to simplify XR rendering
+- Added multipass support for VR
+- Added XR SDK integration (multipass only)
 
 ### Fixed
 - Fixed deserialization crash at runtime